Better input parameters
