+++ coconut/src/lib.rs
@@ -1,6 +1,7 @@
 mod error;
 pub mod issuance;
 pub mod lagrange;
+pub mod proof;
 pub mod request;
 pub mod setup;
