+++ coconut/src/tests/fixtures.rs
@@ -1,6 +1,6 @@
 use crate::{
     issuance::Coin,
-    request::CoinsRequest,
+    request::{CoinsRequest, InputAttribute, OutputAttribute},
     setup::{KeyPair, Parameters, PublicKey},
 };
 use bls12_381::Scalar;
@@ -35,45 +35,70 @@ pub fn aggregated_key() -> PublicKey {
     key
 }
 
+pub fn input_attribute1() -> InputAttribute {
+    InputAttribute {
+        value: Scalar::one(),
+        id: Scalar::from(1234),
+    }
+}
+
+pub fn input_attribute2() -> InputAttribute {
+    InputAttribute {
+        value: Scalar::from(3),
+        id: Scalar::from(5678),
+    }
+}
+
+// Fixture
+pub fn input_attributes() -> Vec<InputAttribute> {
+    vec![input_attribute1(), input_attribute2()]
+}
+
+// Fixture
+pub fn output_attributes() -> Vec<OutputAttribute> {
+    vec![
+        OutputAttribute {
+            value: Scalar::from(2),
+            value_blinding_factor: Scalar::from(10),
+            id: Scalar::from(9123),
+            id_blinding_factor: Scalar::from(20),
+        },
+        OutputAttribute {
+            value: Scalar::from(2),
+            value_blinding_factor: Scalar::from(30),
+            id: Scalar::from(4567),
+            id_blinding_factor: Scalar::from(40),
+        },
+    ]
+}
+
 // Fixture
 pub fn coin1() -> Coin {
-    let value = Scalar::one();
-    let id = Scalar::from(1234);
-    Coin::default(&mut parameters(), &keypair().secret, &value, &id)
+    Coin::default(
+        &mut parameters(),
+        &keypair().secret,
+        &input_attribute1().value,
+        &input_attribute1().id,
+    )
 }
 
 // Fixture
 pub fn coin2() -> Coin {
-    let value = Scalar::from(3);
-    let id = Scalar::from(5678);
-    Coin::default(&mut parameters(), &keypair().secret, &value, &id)
+    Coin::default(
+        &mut parameters(),
+        &keypair().secret,
+        &input_attribute2().value,
+        &input_attribute2().id,
+    )
 }
 
 // Fixture
-pub fn request() -> (CoinsRequest, Vec<(Scalar, Scalar)>) {
-    let public_key = keypair().public;
-    let sigmas = vec![coin1(), coin2()];
-    let input_attributes = vec![
-        (Scalar::from(1), Scalar::from(1234)),
-        (Scalar::from(3), Scalar::from(5678)),
-    ];
-    let output_attributes = vec![
-        (Scalar::from(2), Scalar::from(9123)),
-        (Scalar::from(2), Scalar::from(4567)),
-    ];
-    let blinding_factors = vec![
-        (Scalar::from(10), Scalar::from(20)),
-        (Scalar::from(30), Scalar::from(40)),
-    ];
-
-    let request = CoinsRequest::new(
+pub fn request() -> CoinsRequest {
+    CoinsRequest::new(
         &mut parameters(),
-        &public_key,
-        &sigmas,
-        &input_attributes,
-        &output_attributes,
-        &blinding_factors,
-    );
-
-    (request, blinding_factors)
+        &keypair().public,
+        /* sigmas */ &vec![coin1(), coin2()],
+        &input_attributes(),
+        &output_attributes(),
+    )
 }
